requirements.txt
google-cloud-storage
pyyaml
pandas
pyarrow